fixed all the unit tests
